getting rid of modeller-app class, keep sandbox/scratchPad code from there
